requirements_dev.txt
# dev minimal
numpy>=1.26
pandas>=2.1
matplotlib>=3.8
scipy>=1.11
pyarrow>=14.0
gensim>=4.3
tqdm>=4.66

# prod
fastapi>=0.110
uvicorn>=0.29
mlflow>=2.9
evidently>=0.4

# tools
jupyter

# commands
invoke==2.2.0

# pre-commit
pre-commit==4.1.0

# testing
pytest==8.3.4
coverage==7.8.0

# linting
ruff==0.11.6

# docs
mkdocs==1.6.1
mkdocs-material==9.6.1
mkdocstrings-python==1.16.10

duckdb
faiss-cpu